NEW CMS Analytics
Not using Sam's tagmanager module because it doesn't apply in the CMS,
and there's no straightforward way to add this.

Since this only applies in the CMS, I don't expect it to collide
with any custom Google Analytics setup through the tagmanager module on the website itself.
I'm planning to set up the same Tracking ID there, but via the module (in the master prod database)